+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enhancements
     - Default upstream database version is now `6.4.0`
+    - Improved mobile app detection
     - Upstream improvements for ARM platform detection have been integrated
     - Upstream improvements for Blink engine version detection have been integrated
     - Upstream improvements for browser version detection have been integrated